Upgrades, updates
Fixed some bugs and upgraded module -- getPathTo will now return the actual path to a file, and there is a detectUser function to find a single user. The loadLoginusers method now adds the accountID and id64 to each user.  The examples/switch_user script now works, and auto_cat is mostly functional (needs to actually add the categories).