Fixes, updates, upgrades
Changed loadRegistryLM to be a part of a function named loadRegistry which will load either the Windows registry or registry.vdf on Linux/Mac.

Changed setUser to detect based on the registry, and fallback if there's only one user or error if there's more than one user or no users.

Fixed cat_man to handle partial mode arguments (b or r instead of backup or restore), and updated to use new setUser and loadRegistry instead of loadRegistryLM. It also now uses the 'data' directory like everything else.

Finished auto_cat -- it can now add *or remove* the categories developer, publisher, metacritic score (as a number), No Metacritic, or the N most popular tags where n is either 3 or a user-set argument.

Ignoring file catbackup.json as that was not meant to be included in the first place.