Updoots
Added convertData function to bvdf to auto-convert arrays, timestamps, and booleans from objects, numbers, and 0 or 1 to their internal JS values/string versions of them.

Updated log_info with a saveCache function which will cause it to save some data to the ./examples/data/ cache folder (appinfo1 & 2 an shortcuts1 & 2 each using bvdf and the brt's node-binary-vdf and packageinfo using bvdf).