Bugfixes, additions, and changes. Oh, my.
Fixed Some bugs for loading data that isn't always initalized (shortcuts.vdf, some parts of localconfig, the value of parts of the SteamConfig.user object).

Also fixed a bug in requestOwnedApps where if the user only owns 1 thing it parses a single anonymous Object value and returns it -- it now instead properly returns an Array of a single anon Object.

+ SC.setRoot, SC.setUser methods, and a saveRegistry function (mostly the same as old index.js version).

Working on a half-baked SteamSetting class for easily managing the settings... IDK about all that though -- would need a **very** large `switch` to make it work properly, or some way to target the values by a path so it can play with passed-in objects and return them. So far it has the most useful settings values from registry.vdf, loginusers.vdf, and config.vdf. Pretty nasty...needs help; also untested yet =\