More stuff
Added save-able Text VDF files to save function's handling, some argument error handling for setUser, and disabled packageinfo in SteamPaths -- it's not even useful ATM anyways, so no point in dealing with it.

Renamed /examples/test2.js to test_index2.js as that's the verison of SteamConfig it is for and having two files named test2.js was slightly confusing.

Implemented tests for setRoot and setUser, and added a test for loading all of the data at once.

steam-settings.js is almost finished, but still untested.

Fixed a bug in bvdf.js causing it to fail to parse appinfo properly.